feat(instance): move release and lock responsability to instances (#32)
Moves the responsability to lock and release to the instances themselves. This allows instances to be independent in managing their state and allows for more complex instances

BREAKING CHANGE: the Instance interface has been changed. If you were implementing it directly, you can keep the previous behavior by extending `BasicInstance`